Move constants inside download module
