+++ vrplib/download/download_utils.py
@@ -1,6 +1,6 @@
 import re
 
-from vrplib.constants import CVRP_SETS, DIMACS_NAMES, XXL_NAMES
+from .constants import CVRP_SETS, DIMACS_NAMES, XXL_NAMES
 
 
 def find_set(instance_name: str) -> str:
